==> tcs/1.Data Preprocessing.ipynb <==
# Contents:

### 1.Data Quality Checks
### 2.Creating Dictionary for Treating Duplicate Skills
### 3.Creating DataFrame to replace with dictionary

import pandas as pd
import numpy as np

project = pd.read_excel('S:/Hackathons/TCS/datasets/project.xlsx')
submission = pd.read_excel('S:/Hackathons/TCS/datasets/submission.xlsx')
test = pd.read_excel('S:/Hackathons/TCS/datasets/test.xlsx')
train = pd.read_excel('S:/Hackathons/TCS/datasets/train.xlsx')

project.head()

test.head()

train.head()

## 1.Data Quality Checks

#### 1.Unique Skills Assessment

#### 2.Null Values

# function that returns unique skills when we pass skills column



def make_unique(col):
    all_skills = []
    for i in list(col):
        i = str(i)
        for j in i.split(','):
            all_skills.append(j)
    
    temp = []
    for k in all_skills:
        if k == np.nan:
           print(k)
        else:
            temp.append(k)
        all_skills = temp
        
    
    all_skills = [str(j) for j in all_skills]
    all_skills = [j.lower() for j in all_skills]
    all_skills = [str(j) for j in all_skills]
    all_skills = [j.strip() for j in all_skills]
    all_skills = [j.replace('.' , '') for j in all_skills]
    unique_skills = list(set(all_skills))
    
    return unique_skills


# concating all unique skills of train,test,project

p = make_unique(project['skills'])
q = make_unique(train['skills']) 
r = make_unique(test['skills'])


unique_skills = list(set(p+q+r))
unique_skills.sort()

unique_skills = pd.DataFrame(unique_skills,columns = ['skill'])
unique_skills.dropna(inplace = True)

unique_skills.head(25)

#### TakeAways

1.from above we could see that there were really many duplicates like (ai, ai/ml, ai/ ml)

2.we need to clean the data before we build the model

3.It could be done by either regex or manually ,  i preffered manually

## 2.Creating Dictionary for Treating Duplicate Skills

#these are rules for replaceing texts
rules = {

 'adobe familiarity with photoshop' : 'adobe photoshop',

 'advanced data structures and algorithms': 'dsa',
 'ai' : 'ai/ml',
 'ai  ml  deep learning' : 'ai/ml',
 'ai / ml' : 'ai/ml',
 'ai/ ml' : 'ai/ml',
 'ai/ml' :'ai/ml',
 'ai/ml basics' : 'ai/ml',



 'algorithm' : 'daa',
 'algorithm design and analysis' : 'daa',
 'algorithms'  : 'daa',
 'algorithms and analysis design'  : 'daa',
 'algorithms and data structure'  : 'dsa',
 'algorithms and data structures'  : 'dsa',
 'algorithms and datastructures'  : 'dsa',

 'alibaba cloud clouder professional' :  'alibaba cloud',
 'alogorithms': 'daa',
 'amazon   aws' : 'aws',
 'amazon web services' : 'aws',
 
 'aml analytics' : 'ai/ml',
 'analytical skills' : 'analytics',

 'android' : 'android app developement',

 'android app developer' : 'android app developement',
 'android app development' : 'android app developement',
 'android application development' : 'android app developement',
 'android developmen' : 'android app developement',
 'android development' : 'android app developement',
 'android development using kotlin' : 'kotlin',

 'angular  based hybrid mobile app' : 'angular',
 'angular 6' : 'angular',
 'angular 6+': 'angular',
 'angular 7' : 'angular',
 'angular 7  framework' : 'angular',
 'angular 7+': 'angular',
 'angular js' : 'angular',
 'angular js/ node js' : 'angular',
 'angular/js' : 'angular',
 'angular6' : 'angular',
 'angular7' : 'angular',
 'angular8' : 'angular',
 'angularjs' : 'angular',
 'angural js' : 'angular',
 
 
 'ansys electronics' : 'ansys',
 'ansys fluent' : 'ansys',

 'any js framework (knockout js/angular js)' : 'javascript',
 'any open source bi tool' : 'tablea',
 'any programming skills' : 'programming',

 'api development' : 'api development and testing',
 'api testing  postman': 'api development and testing',
 'app developement': 'app development',
 'application designing': 'app development',
 'application development': 'app development',
 'apps': 'app development',
 'ar' : 'ar/vr',
 'ar/vr programming' : 'ar/vr',
 'arcore' : 'ar/vr',
 
 'arduino coding':'arduino',
 'arduino controller':'arduino',
 'arduino programing':'arduino',
 'arduino programming':'arduino',
 'arduino uno':'arduino',
 
 'artificail intelligence' : 'ai/ml',
 'artifical intellegence' : 'ai/ml',
 'artificial intelligence' : 'ai/ml',
 'artificial intelligence   ml  dl' : 'ai/ml',
 'artificial intelligence and machine learning'  : 'ai/ml',
 'artificial intelligence deep learning'  : 'ai/ml',
 'artificial intelligence techniques' : 'ai/ml',
 
 
 'aspnet   mvc':'aspnet',
 'assembly 8086 and 8051' :'assembly',
 'assembly language' :'assembly',
 'augmented reality' : 'ar/vr',
 'auto cad' : 'autocad',
 
 
 'autodesk maya/blender' : 'autodesk maya',

 'automation and robotics' : 'automation',
 'automation studio' : 'automation',
 'automobile basics': 'automation',


 'aws cloud practitioner': 'aws',
 'aws ec2': 'aws',
 'aws machine learning': 'aws',
 'azur': 'azure',
 
 'azure chatbot service': 'azure',
 'azure cloud basics': 'azure',
 'azure"': 'azure',
 'azurecli': 'azurecli',
 'azurecli and bash scripting' : 'azurecli',
 
 
 'bash scripting' : 'bash',
 'basic cloud skills' : 'cloud',
 'basic coding knowledge on java' : 'java',
 'basic concept of blockchain' : 'blockchain',
 'basic data science with python' : 'ai/ml',
 'basic idea of image and video analytic' : 'cv',
 'basic java understanding' : 'java',
 'basic knowledge of php' : 'php',
 
 'basic ml' : 'ai/ml',
 'basic programming knowledge' : 'programming',
 'basic programming skills' : 'programming',
 'basics of blockchain': 'blockchain',
 'basics of clustering and other algorithm' : 'ai/ml',
 'basics of db' : 'dbms',
 'basics of iot' : 'iot',
 'basics of python' : 'python',

 
 'big data analysis' : 'big data',
 'big data analytics': 'big data',
 'big data processing': 'big data',
 'big data tools': 'big data',

 'blender 3d' : 'blender',
 'block chain' : 'blockchain',
 'blockchain' : 'blockchain',
 'blockchain fundamentals' : 'blockchain',
 'blockchain technology' : 'blockchain',

 'bootstrap 4' : 'bootstrap',
 'bootstrap3': 'bootstrap',
 
 
 'c' : 'c++',
 'c  programming language''c' : 'c++',
 'c & sql''c' : 'c++',
 'c ++''c' : 'c++',
 'c and c++''c' : 'c++',
 'c and cpp''c' : 'c++',
 'c c++''c' : 'c++',
 'c cpp programming''c' : 'c++',
 'c graphics''c' : 'c++',
 'c language''c' : 'c++',
 'c prog''c' : 'c++',
 'c programming''c' : 'c++',
 'c programming language''c' : 'c++',
 'c sharp''c' : 'c++',
 'c sharp scripting''c' : 'c++',
 'c#''c' : 'c++',
 'c++''c' : 'c++',
 'c++  c''c' : 'c++',
 'c++ c''c' : 'c++',
 'c++ language''c' : 'c++',
 'c++ programming''c' : 'c++',
 'c++ programming language''c' : 'c++',
 'c++ python''c' : 'c++',
 'c++14''c' : 'c++',
 
'c' : 'c++',
'c   c++': 'c++',
'c  c++': 'c++',
'c  c++  java': 'c++',
'c  programming language': 'c++',
'c & sql': 'c++',
'c ++': 'c++',
'c and c++': 'c++',
'c and cpp': 'c++',
'c c++': 'c++',
'c cpp programming': 'c++',
'c graphics': 'c++',
'c language': 'c++',
'c prog': 'c++',
'c programming': 'c++',
'c programming language': 'c++',
'c sharp': 'c#',
'c sharp scripting': 'c#',
'c#': 'c#',
'c++  c': 'c++',
'c++ c': 'c++',
'c++ language': 'c++',
'c++ programming': 'c++',
'c++ programming language': 'c++',
'c++ python': 'c++',
'c++14': 'c++',
 
 'cassandradb' : 'cassandra',
 'cloud computing' : 'cloud',
 'cloud computing   aws' : 'aws',
 'cloud computing  aws' : 'aws',
 'cloud computing  gcp' : 'gcp',
 'cloud concepts': 'cloud',
 'cloud configuration & management': 'cloud',
 'cloud knowledge': 'cloud',

 'cn': 'computer networking',
 'cnn': 'cv',
 'coding' : 'programming',
 'coding skills' : 'programming',
 'communication - verbal & written':'aptitude',
 'competitive coding': 'programming',
 'competitive programming': 'programming',
 
 'computer networks' : 'computer networking',
 'computer vision' : 'cv',
 'concepts on cloud and knowledge about azure' : 'cloud',
 'containerisation using docker' : 'docker',
 'converting wireframes/ visual designs into ui' : 'ux/ui',
 'convolutional neural networks': 'cv',
 'core blockchain' : 'blockchain',
 'core java' : 'java',
 'core php':'php',
 'cpp' : 'c++',
 'cpp programming' : 'c++',
 'csharp':'c#',
 'css' : 'css3',
 
 'css 3( bootstrap)' : 'css3',
 'css3' : 'css3',
 
 'cuda programming' : 'cuda',

 'data analysis': 'ai/ml',
 'data analysis and visualisation  r  python  octave':'ai/ml in r',
 'data analytics': 'ai/ml',
 'data analytics   machine learning': 'ai/ml',
 'data analytics using python': 'ai/ml',
 'data base' : 'dbms',
 'data base management system' : 'dbms',
 'data base management systems' : 'dbms',
 'data modeling':'data modelling',
 
 'data science': 'ai/ml',
 'data science and analytics' : 'ai/ml',
 'data science and analytics using r': 'ai/ml in r',
 'data scientist': 'dsa',
 'data strucres': 'dsa',
 'data structer' : 'dsa',
 'data structure': 'dsa',
 'data structure   algorithm': 'dsa',
 'data structure   algorithms': 'dsa',
 'data structure and algorithm': 'dsa',
 'data structure and algorithm design': 'dsa',
 'data structure and algorithms': 'dsa',
  'ds': 'dsa',
 'ds and algo': 'dsa',

 'data structures': 'dsa',
 'data structures   algorithms': 'dsa',
 'data structures and algorithm': 'dsa',
 'data structures and algorithms': 'dsa',
 'data structures and object oriented concepts': 'dsa',
 'data structutres': 'dsa',
 'data visualisation' : 'visualization',
 'data visualization' : 'visualization',
 
 'data warehousing knowledge' : 'data warehouse',
 'database': 'dbms',
 'database exposure': 'dbms',
 'database management' : 'dbms',
 'database management system' : 'dbms',
 'database management systems' : 'dbms',
 'database system' : 'dbms',
 'databases' : 'dbms',
 'datastructures': 'dsa',
 'datastructures and algorithms': 'dsa',
 'db concepts': 'dbms',
 
 'dbms  mysql': 'dbms',
 'dbms sql': 'dbms',
 'dbms using sql': 'dbms',
 
 'debugging skill' : 'debugging',
 'deep learing' : 'deep learning',
 'deep learning and neural networks' : 'deep learning',
 'deep learning python' : 'deep learning',
 'deep programming skills' : 'programming',
 'deeplearning': 'deep learning',
 'design  and analysis of algorithm':'daa',
 'design and analysis of algorithms':'daa',
 'design and product development   innovation' : 'product development',
 'desirable - knowledge of cloud infrastructure such as aws': 'aws',
 'dev ops': 'devops',
 'development and devops': 'devops',
 'development skills': 'devops',

 'devops tools' : 'devops',
 'digital circuits' : 'electronics circuits',
 'digital electronics' : 'electronics circuits',
 'digital security' : 'security',
 'digital signal processing' : 'signal processing',

 'digital skills having programming language' : 'programming',
 'django': 'django',
 'django framework' : 'django',
 'django rest framework': 'django',
 
 'dockers' : 'docker',
 'dot net' : 'dotnet',
 
 'drupal and in dept knowledge of js' : 'javascript',

 
 'elastic/solar for search index' :'elasticsearch' ,
 'elastic search' : 'elasticsearch',
 
 'elk stack' : 'elk',
 
 'embedded c arduino programming':'embedded c',
 'embedded systems':'embedded c',
 
 'ethereum solidity programming' : 'ethereum',
 'ethical hacking' : 'hacking',
 'ethical hacking basic' : 'hacking',

 'experience in sql and python scripting' : 'sql',
 'exposure to ai/nlp in coursework':'ai/ml',
 'exposure to coding':'programming',
 'exposure to machine learning concepts and algorithms' : 'ai/ml',
 'exposure to tableau or any other visualization tool can be an added advantage' : 'tableau',
 'express api' : 'express',
 'express js': 'express',
 'expressjs': 'express',
 'firebase database' : 'firebase',
 'firestore' : 'firebase',
 'flask micro web framework' :  'flask',
 'flimora': 'adobe photoshop',
 
 'flutter   dart' :  'flutter',
 'flutter  mobile app development':  'flutter',
 'flutter framework':  'flutter',
 'framework  django flask nodejs expressjs' : 'flask',
 'front end  development' : 'front end  development',
 'front end technologies': 'frontend developement',
 'front end technologies html css js': 'frontend developement',
 'front end web designer': 'frontend developement',
 'front end web development' : 'frontend developement',
 
 
 'full stack developer' : 'full stack web development',
 'full stack developer front end server side' : 'full stack web development',
 'full stack development' : 'full stack web development',
 'full stack web developement' : 'full stack web development',
 'full stack web development' : 'full stack web development',
 'full-stack programing skills' : 'full stack web development',
 'fundamental of cyber security' : 'cyber security',
 'fundamentals of blockchain' : 'blockchain',

 
 
 'git' : 'git',
 'git github' : 'git',
 'git version control' : 'git',
 'github': 'git',
 'gitlab': 'git',
 'gnu radio': 'gnuradio',

 'go lang' : 'go',
 'golang': 'go',
 'good analytical skills' : 'analytical',
 'good aptitude':'aptitude' ,
 'good coding skills':'programming',
 'good programming skills' : 'programming',
 'good rdbms understaning and pl-sql expertise' : 'dbms',
 'good understanding of  android platform(raspberry pi 3)' : 'ar/vr' ,
 'good understanding of ar/vr concepts' : 'ar/vr',
 'graphic designing' :  'graphics designing',


 'hadoop framework' : 'hadoop',
 'handons experience on opencv' : 'cv',
 'hiberante' : 'hibernate',

 'html': 'web development',
 'html   css': 'web development',
 'html   css   js': 'web development',
 'html  css': 'web development',
 'html  css   js': 'web development',
 'html  css  javascript': 'web development',
 'html  css  js': 'web development',
 'html  css  js  php  angular js  nodejs': 'web development',
 'html  css and js': 'web development',
 'html 5': 'web development',
 'html 5 and css3': 'web development',
 'html and css': 'web development',
 'html css': 'web development',
 'html css bootstrap': 'web development',
 'html css javascript': 'web development',
 'html css js microservices': 'web development',
 'html css js php': 'web development',
 'html+css': 'web development',
 'html+css  js': 'web development',
 'html+css+js+django': 'web development',
 'html5': 'web development',
 'html5   css3' : 'web development',
 
 
 
 
 'illustartor' : 'illustrator',
 'image   video processing': 'cv',
 'image processing' : 'cv',
 'image processing and computer vision' : 'cv',
 'implementation of data pipe line and search indexes' : 'data infa',
 'intend classification and image processing' : 'cv',
 'internet of things' : 'iot',
 'internet of things   iot' : 'iot',
 'ionic framework': 'ionic',
 'ios and android development' : 'ios',
 'ios development' : 'ios',
 'iot and device design' : 'iot',
 'iot technologies' : 'iot',
 'iots' : 'iot',


 'java  core+swing framework' :'java',
 'java advanced' :'java',
 'java and object oriented programming' :'java',
 'java basic knowledge' :'java',
 
 
 
 
 'ui design':  'ui/ux',
 'ui path' : 'uipath',
 'ui technologies (angular)':  'ui/ux',
 'ui ux':  'ui/ux',
 'ui ux design':  'ui/ux',
 'ui/ux' :  'ui/ux',
 'ui/ux designing skills' :  'ui/ux',
 'understanding of angular framework' : 'angular',
  'understanding of ci/cd' : 'ci/cd',
 'unity 3d' : 'unity',
 'unity 5 game engine' : 'unity',
 'unity development' : 'unity',
 'unity game development' : 'unity',
 'unity pro' : 'unity',
 'unity3d' : 'unity',
 'unix commands' : 'unix',
 'unix programming' : 'unix',


 
 'web api development' : 'web api',
 'web application development' : 'web api',
 'web design' :'web design',
 'web desining' : 'web design',
 'web develepment' : 'web development',
 'web developement' : 'web development',
 'web development' : 'web development',
 'web devloping' : 'web development',
 'web programming' : 'web development',
'web technologies' : 'web technology',
 
 
 'java  core+swing framework':'java',
 'java advanced':'java',
 'java and object oriented programming':'java',
 'java basic knowledge':'java',
 'java core':'java',
 'java ee':'java',
 'java jee':'java jse',

 'java language':'java',
 'java programming':'java',
 'java script':'javascript',
 'java scripting knowledge':'javascript',
 'java se':'java',
 'java spring boot': 'java springboot',

 'java sprint boot': 'java springboot',
 'java/j2ee':'javascript',

 'javacsript':'javascript',

 'javascprit':'javascript',
 
 'javascript   es6   programming language':'javascript',
 'javascript   jquery':'jquery',
 'javascript nodejs  express':'javascript',

 
 'js':'javascript',
 'js full stack development':'javascript',
 
 'junit5':'junit',

 'knowledge of mobile app development':'mobile app development',
 'knowledge on libraries such as opencv':'cv',
 'knowledge on public blockchain networks such as bitcoin':'blockchain',

 'lab view' : 'labview',



 'linux rhel':'linux',
 'linux shell scripting':'linux',
 'linux shell utilities':'linux',

 'machine leanring' : 'ai/ml',
 'machine learning': 'ai/ml',
 'machine learning & artificial intelligence': 'ai/ml',
 'machine learning algorithms': 'ai/ml',
 'machine learning and ai': 'ai/ml',
 'machine learning basics': 'ai/ml',
 'machine learning for text summarization': 'ai/ml',
 'machine learning tools & framework': 'ai/ml',
 'machine learning using python': 'ai/ml',
 'machine learning with python': 'ai/ml',

 'materialize css': 'css',
 'mathematics' :  'mathematical modeling',

 'mean' :  'mean stack',

 'mechine learnig': 'ai/ml',
 'micro services' : 'micro service architecture',
 'microcontrollers  arduino  nodemcu': 'arduino',
 'microsoft  excel' : 'microsoft office',
 'microsoft  excel word powerpoint' : 'microsoft office',
 'microsoft azure':'azure',
 'minimal knowledge on cloud computing -minimal knowledge on backup' : 'cloud',

 'ml': 'ai/ml',
 'ml & python': 'ai/ml',
 'ml /ai': 'ai/ml',
 'ml ai': 'ai/ml',
 'ml and dl': 'ai/ml',
 'ml development knowledge': 'ai/ml',
 'ml techniques': 'ai/ml',
 'ml/dl': 'ai/ml',
 
 
 'mobile development skills':'mobile app development',
 'modern web development full stack':'modern web development',
 'mondo db': 'mongodb',
 'mongo db' : 'mongodb',
 

 'mongoose os': 'mongodb',

 'ms excel':'msoffice',
 'ms office':'msoffice',
 'ms powerpoint':'msoffice',
 'ms sql':'sql',
 'ms word':'msoffice',
 'my sql':'sql',
 'mysql':'sql',
 'mysql  oracle sql':'oracle',
 'mysql - good database concepts':'dbms',
 'mysql oracle':'oracle',

 'natural language processing':'nlp',
 'natural language programming':'nlp',


 'network building and simulation':  'net concepts',
 'network security': 'computer networking',
 'network securityapi integration': 'computer networking',
 'networking': 'computer networking',
 'networking  ccna': 'computer networking',
 'neural network':'ann',
 'neural networks':'ann',
 'neural networks and deep learning':'deep learning',
 
 

 'object oriented programing':  'oops',
 'object oriented programming':  'oops',
 'object oriented programming concept':  'oops',
 'objected oriented programming':  'oops',

 'oop concepts' :  'oops',

 'open cv':'cv',

 'opencv':'cv',

 'operating system' :'os',
 'operating systems':'os',

 'oracle 10g dbms': 'oracle',

 'pattern recognition':'ai/ml',
 'pega':'pega tool',
 'perl' : 'perl& shell',
 'perl& shellapi integration' : 'perl& shell',



 'php basics':'php',


 'postgis': 'postgresql',
 'postgres': 'postgresql',
 'postgres database': 'postgresql',

 'power bi':'bi',
 
 'problem solving using programming':'problem solving',
 'programming in c':'c',
 'programming in c language':'c',
 'programming in c++':'c++',
 'programming in java':'java',
 'programming in python':'python',

 
 'programming using c':'c',
 'programming using python':'python',
 'programming with c++':'c++',
 'programming with java': 'java',
 'programming with python' : 'python',

 
 'pyhon' : 'python',
 'pyhthon' : 'python',
 'pyhton' : 'python',
 'pyhton3' : 'python',
 'pyqt' : 'python',
 'python' : 'python',
 'python  programming skills' : 'python',
 'python 3' : 'python',
 'python 30' : 'python',
 'python 3x' : 'python',
 'python for machine learning': 'python libraries',
 'python frameworks  django  flask' : 'flask',

 'python programming': 'python',
 'python/r programming': 'python',
 'python3': 'python',


 
 'r' :  'r programming',

 'r programming language' :  'r programming',

 'rapid miner  tableau':'rapid miner',


 'rpa with ai & ml':'ai/ml',

'r' : 'r programming',
 'r programming language'  :  'r programming',
 'react js' : 'reactjs',
 'react library' : 'reactjs',
 'react native' : 'reactnative',
 'rest api development' :  'rest api',
 'rest webservices' : 'rest services',


 'shell programing' : 'shell scripting',
  'solid works': 'solidworks',
   'sound knowledge of dbms' : 'dbms',
 'sound knowledge of management fundamentals as per current syllabus' : 'management',
 'spring & hibernate' : 'spring hibernate',
  'statictics':'statistics',
  'structured and unstructured database' : 'dbms',
   
 'tensor flow' :'tensorflow' ,
 'tenserflow' : 'tensorflow',



 'ui design':  'ui/ux',
 'ui path' : 'uipath',
 'ui technologies (angular)':  'ui/ux',
 'ui ux':  'ui/ux',
 'ui ux design':  'ui/ux',
 'ui/ux' :  'ui/ux',
 'ui/ux designing skills' :  'ui/ux',
    
'understanding of angular framework' : 'angular',
 'understanding of ci/cd' : 'ci/cd',
 'unity 3d' : 'unity',
 'unity 5 game engine' : 'unity',
 'unity development' : 'unity',
 'unity game development' : 'unity',
 'unity pro' : 'unity',
 'unity3d' : 'unity',
 'unix commands' : 'unix',
 'unix programming' : 'unix',


 
 'web api development' : 'web api',
 'web application development' : 'web api',
 'web design' :'web design',
 'web desining' : 'web design',
 'web develepment' : 'web development',
 'web developement' : 'web development',
 'web development' : 'web development',
 'web devloping' : 'web development',
 'web programming' : 'web development',
'web technologies' : 'web technology',

}


.
.
.



Now we have good Dictionary to replace all duplicates.

The best way of replacing all duplicates when we have dictionary is to  create a Dataframe that consists of candidateid and skill as columns




.
.
.

# function that creates dataframe of columns id and skill when we pass id and sring of skills


def make_df_for_row(candidate,skills):
    
    skills = str(skills)
    skills = list(skills.split(','))
    
    
    #removes nan
    temp = []
    for k in skills:
        if k == np.nan:
           print(k)
        else:
            temp.append(k)
        skills = temp
     
    
    
    skills = [str(j) for j in skills]
    skills = [j.lower() for j in skills]
    skills = [str(j) for j in skills]
    skills = [j.strip() for j in skills]
    skills = [j.replace('.' , '') for j in skills]
    
    temp_dict = []
    for skill in skills:
        
        temp_dict.append({
            
            'name': candidate,
            'skill': skill
            
            })
        
    return pd.DataFrame(temp_dict)

# For Example

make_df_for_row(train['can_id'][0],train['skills'][0])

## 3.Creating DataFrame to replace with dictionary

# Now we have a function for just 1 row

#for one go, I have created another function that creates dataframe for multiple rows using above function

def make_df_multiple_rows(col1,col2):
    
    main_df = pd.DataFrame()
    for i,j in zip(list(col1),list(col2)):
    
        temp_df = make_df_for_row(i,j)
        main_df = pd.concat([main_df,temp_df])
        
    return main_df

train_id_skill = make_df_multiple_rows(train['can_id'],train['skills'])

# similarly for test and project 
test_id_skill = make_df_multiple_rows(test['can_id'],test['skills'])
proj_id_skill = make_df_multiple_rows(project['pro_id'],project['skills'])

train_id_skill.head()

# replace train,test and project with rules dictionary
train_id_skill_replaced = train_id_skill.replace(rules)
test_id_skill_replaced = test_id_skill.replace(rules)
proj_id_skill_replaced = proj_id_skill.replace(rules)

# concat train,test,proj before replacing 

temp = pd.concat([train_id_skill,test_id_skill])
temp =  pd.concat([proj_id_skill,temp])

# concat train,test,proj after replacing 


temp1 = pd.concat([train_id_skill_replaced,test_id_skill_replaced])
temp12 =  pd.concat([proj_id_skill_replaced,temp1])

print(temp1.head(),temp1.shape)

print('Total Number of Unique skills acorss train,test,project datasets before replacing are :',temp['skill'].nunique())
print('Total Number of Unique skills acorss train,test,project datasets after replacing are :',temp1['skill'].nunique())

skills_list = pd.DataFrame(temp12.groupby(['name'])['skill'].apply(lambda x: ' '.join(x))).reset_index()

#temp12.to_csv('S:/Hackathons/TCS/datasets/name_skills_df1.csv',index = False)

temp12.shape

k.shape


k.drop_duplicates(inplace = True)
k_ct = pd.crosstab(k['name'], k['skill'],margins = True)
k_ct.shape

skills_list
