==> tcs/2.Feature Engineering Part 1 - Making Super Functions.ipynb <==
# 1.Top_Skills_Score

There will be more chance of getting hired if the candidate has top demanding skills.

Scoring be like:

0.1 - the candidate have 1 out top 10 skills

0.2 - the candidate have 2 out top 10 skills

likewise ............

import pandas as pd
import numpy as np

train = pd.read_csv('S:/Hackathons/TCS/train1.csv')
skills_list = pd.read_csv('S:/Hackathons/TCS/skills_list.csv')
name_skills_df = pd.read_csv('S:/Hackathons/TCS/datasets/name_skills_df1')

name_skills_df.drop_duplicates(inplace = True)

selected_cands = list(train[train['evaluation_status'] == 'SELECTED']['can_id'])


selected_can_inv_skills = name_skills_df[name_skills_df['name'].isin(selected_cands)]
selected_can_ct = pd.crosstab(selected_can_inv_skills['name'], selected_can_inv_skills['skill'],margins = True)


num_skills_sel = pd.DataFrame(index = selected_can_ct.columns)
num_skills_sel['#_sel_skills'] = list(selected_can_ct.loc[['All']].values[0])

num_skills_sel = num_skills_sel.sort_values(by = ['#_sel_skills'],ascending = False)
num_skills_sel.drop('All',inplace =True)

num_skills_sel.head(10)

top_skills = list(num_skills_sel.index[0:11])




id_vs_skills_one_hot = pd.crosstab(name_skills_df['name'], name_skills_df['skill'],margins = True)
id_vs_skills_one_hot.reset_index(inplace = True)

id_vs_skills_one_hot[['name']+top_skills]

top_skills_score = id_vs_skills_one_hot[['name']+top_skills].sum(axis = 1, skipna = True)*0.1
id_vs_skills_one_hot['top_skills_score'] = top_skills_score

top_skills_score = id_vs_skills_one_hot[['name','top_skills_score']]
top_skills_score.head(7)

#top_skills_score.to_csv('S:/Hackathons/TCS/top_skills_score.csv',index = False)

id_vs_skills_one_hot.head(10)


# 2.id lookup 

Both candidate and project variables in train data have high cardinality.
One way of dealing with high cardinality features is by producing LookUp Table

The Lookup's can be produced by many methods like from

1.Embedding Layer of Keras.

2.Encoding from other features of dataset.

3.....

For this case The best way of representing the skills of candidate and project id OneHotEncoding as you see above

Clearly, OnHotEncoded dataframe id_vs_skills_one_hot has high sparsity.We can apply Singular Value Decomposition on this high 
sparsity data and use latent features of candidate_id and project_id as a lookup table


from scipy.sparse.linalg import svds
id_vs_skills_one_hot.index = id_vs_skills_one_hot.name
id_vs_skills_one_hot.drop(columns = ['name'],inplace = True)

matrix = id_vs_skills_one_hot
matrix = matrix.astype(float)
U, sigma, Vt = svds(matrix)

U = pd.DataFrame(U,columns = ['u1','u2','u3','u4','u5','u6'])

U.index = id_vs_skills_one_hot.index
U

#U = U.to_csv('S:/Hackathons/TCS/id_lookup.csv',index = True)
