==> tcs/3.Feature Engineering Part 2 - Creating Train and Test Datasets.ipynb <==
import pandas as pd
import numpy as np

train = pd.read_csv('S:/Hackathons/TCS/train1.csv')
test = pd.read_csv('S:/Hackathons/TCS/test1.csv')
submission = pd.read_excel('S:/Hackathons/TCS/datasets/submission.xlsx')
project = pd.read_csv('S:/Hackathons/TCS/project1.csv')

# 1.same location or not

# Data Prepeocessing for location

train_loc = pd.DataFrame(train[['can_id','location']])
test_loc = pd.DataFrame(test[['can_id','location']])
proj_loc = pd.DataFrame(project[['pro_id','location']])
proj_loc.columns = ['can_id','location']


temp = pd.concat([train_loc,test_loc])
locations = pd.concat([temp,proj_loc])
locations.columns = ['id','location']
locations.drop_duplicates(inplace = True)

locations.index = locations.id
locations.drop(columns = ['id'],inplace = True)
locations['location'].astype(str)


# function

def is_same_loc(c,p):
    if locations.loc[c].values[0] == locations.loc[p].values[0]:
        return 1
    else :
        return 0


is_same_loc('c1','p3')

is_same_loc('c5','p20')

# 2.mode

# preprocess data for mode

proj_loc = project[['pro_id','mode']]
proj_loc.index = proj_loc.pro_id
proj_loc.drop(columns = ['pro_id'],inplace = True)


# function

def mode(p):
    if proj_loc.loc[p].values[0] == 'PHYGITAL':
       return 1
    else:
       return 0


mode('p20') 

mode('p3') 


# 3.number of common skills



skills_list = pd.read_csv('S:/Hackathons/TCS/skills_list.csv')
skills_list.dropna(inplace = True)



def numb_comm_skills(c,p):
    s1 = set(skills_list.loc[skills_list['name'] == c]['skill'].values[0].split(','))
    s2 = set(skills_list.loc[skills_list['name'] == p]['skill'].values[0].split(','))  
    
    return len(list(s1.intersection(s2)))

numb_comm_skills('c20','p200')



# 4.top skills score

There will be more chance of getting hired if the candidate has top demanding skills.

Scoring be like:

0.1 - the candidate have 1 out top 10 skills

0.2 - the candidate have 2 out top 10 skills

likewise ............



top_skills_score = pd.read_csv('S:/Hackathons/TCS/top_skills_score.csv')
top_skills_score.index = top_skills_score.name
top_skills_score.drop(columns = ['name'],inplace = True)


def top_score(c):
    a = top_skills_score.loc[c].values[0]
    return a



top_score('c1')



# 5.id lookup 

First Made a Look up table for candidate and Project is using Matrix Factorization
     
Data used for matrix Factorization is One Hot Encoded Table of both (Candidate_id and Project_id) in index
and skills at columns 



###############################


#5.id lookup 


id_lookup = pd.read_csv('S:/Hackathons/TCS/id_lookup.csv')
id_lookup.index =  id_lookup.name
id_lookup.drop(columns = ['name'],inplace = True)

def create_lookup(c,p):
    a = list(id_lookup.loc[c].values)
    b = list(id_lookup.loc[p].values)
    
    return np.add(a,b)
create_lookup('c20','p200')


# 6.euclidian between the lookup

Euclidian distance between Latent Vectors of project and candidate id

From Matrix Factorization. First I used Latent Vector for each project and user id as an embedding on the table
but didn’t give better result. Then I Used latent vector for Candidate and project as point and computer 
Euclidian distance between them



from scipy.spatial import distance


def create_eulc(c,p):
    a = id_lookup.loc[c].values
    b = id_lookup.loc[p].values
    
    return distance.euclidean(a, b)
create_eulc('c20','p200')

# 7 jacard score

jacard score for finding the similarity between the one hot encoded of Candidate and Project.


id_vs_skills_one_hot = pd.read_csv('S:/Hackathons/TCS/one_hot.csv')   #id_vs_skills_one_hot
id_vs_skills_one_hot.index = id_vs_skills_one_hot.name
id_vs_skills_one_hot.drop(columns = ['name'],inplace = True)
id_vs_skills_one_hot.drop(columns = ['All'],inplace = True)
id_vs_skills_one_hot.drop('All',inplace = True)

from sklearn.metrics import jaccard_score
def get_jacard_scores(c,p):
    return jaccard_score(id_vs_skills_one_hot.loc[c].values, id_vs_skills_one_hot.loc[p].values)

c = 'c20'
p = 'p200'

get_jacard_scores(c,p)


def create_df(c_col,p_col):
    temp_df = []
    for c,p in zip(c_col,p_col):
        row = [is_same_loc(c,p)]+ [mode(p)]+ [numb_comm_skills(c,p)]+[top_score(c)]+[create_eulc(c,p)]+[get_jacard_scores(c,p)]
        temp_df.append(row)
    
    columns = ['is_same_loc','mode','numb_comm_skills','top_score','eulc_dist','jacard_score_b/w_skills']
    
    df = pd.DataFrame(temp_df,columns = columns) 
    
    return df



### Train Data Creation

train.dropna(inplace = True)
train_df = create_df(train['can_id'],train['pro_id'])


train_df['result'] = list(train['evaluation_status'])
#train_df.to_csv('S:/Hackathons/TCS/train_main.csv',index = False)
train_df.head()

test.isnull().sum()

test.dropna(inplace = True)
test_pro = []
test_can = []
for i in submission['Submission ']:
    for j in test['can_id']:
        test_pro.append(i)
        test_can.append(j)
        

test.dropna(inplace = True)
test_main = create_df(test_can,test_pro)


test_main['proj_id'] = test_pro
test_main['cand_id'] = test_can

#test_main.to_csv('S:/Hackathons/TCS/test_main1.csv',index = False)

test_main.head()
