==> tcs/4.Predictive Modeling.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np


train_main = pd.read_csv('S:/Hackathons/TCS/train_main.csv')
train_main.drop(columns = ['proj_id','cand_id'],inplace = True)

train_main.head()


from sklearn.model_selection import train_test_split
X_tr,X_tst,y_tr,y_tst = train_test_split(train_main.drop(columns = ['result']),train_main['result'],
                                         stratify = train_main['result'])

import xgboost

params_xg = {
        'n_estimators': [750,1000,1250,1500,2000],
        'max_depth':[7,15,20],
        'min_child_weight': [3, 10],
        'gamma': [0.5,2, 5],
        'subsample': [0.6, 0.8],
        'colsample_bytree': [0.7, 1.0]
        }


from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error,log_loss,accuracy_score
score_card = pd.Series()

def best_grid_model(estimator,X_tr,X_tst,y_tr,y_tst,parmas_grid):
    
    grid = GridSearchCV(estimator=estimator,
                    param_grid=parmas_grid,
                    cv=3,  
                    verbose=10, 
                    n_jobs=-1,
                    scoring='accuracy',
                    refit = True)
    
    grid.fit(X_tr,y_tr)
    print('Best Parmaters are :',grid.best_params_)
    print('ReFitting Best Parameters to New Estimator ...............','\n')
    print('Best Score is :',grid.best_score_)
    best_estimator = estimator
    best_estimator.set_params(**grid.best_params_)
    best_estimator.fit(X_tr,y_tr)
    
    
    best_accuracy_score = accuracy_score(y_tst,best_estimator.predict(X_tst))
    print('RMSE for X_test is :',best_accuracy_score,'\n' )
    
    score_card['best_{}'.format(estimator)] = best_accuracy_score
    
    return grid,best_estimator

grid_xgb,best_xgb = best_grid_model(xgboost.XGBClassifier(),X_tr,X_tst,y_tr,y_tst,params_xg)

test_main = pd.read_csv('S:/Hackathons/TCS/test_main.csv')
best_xgb.predict_proba(test_main)


best_xgb.classes_

preds = pd.DataFrame(best_xgb.predict_proba(test_main),columns = best_xgb.classes_)
preds.to_csv('S:/Hackathons/TCS/preds1.csv',index = False)
