==> tcs/5.Results Preprocessing.ipynb <==
import pandas as pd

test_main = pd.read_csv('S:/Hackathons/TCS/test_main.csv')
test_main1 = pd.read_csv('S:/Hackathons/TCS/test_main1.csv')


preds = pd.read_csv('S:/Hackathons/TCS/preds1.csv')
preds['proj_id'] = test_main1['proj_id']
preds['cand_id'] = test_main1['cand_id'] 

preds

SELECTED Column consists of Probability of Selection

df_agg = preds.groupby(['proj_id','cand_id']).agg({'SELECTED':sum})
df_agg.shape

df_agg.head(5)

df_agg.tail(5)

grps = df_agg['SELECTED'].groupby('proj_id', group_keys=False)
grps.head()

top5s = grps.apply(lambda x: x.sort_values(ascending=False).head(5))
top5s

#top5s.to_csv('S:/Hackathons/TCS/preds.csv',index = False)
top5s1 =top5s.copy()

top5s1 = pd.DataFrame(top5s1)
top5s1.reset_index(inplace = True)
top5s1.drop(columns = ['SELECTED'],inplace = True)

submission = pd.DataFrame(top5s1.groupby(['proj_id'])['cand_id'].apply(lambda x: ','.join(x))).reset_index()

submission

submission.shape

#final.to_csv('S:/Hackathons/TCS/final.csv',index = False)
